update: removed unnecessary fileds in the request format
- separated new features (interest, about) route handler.
